% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click_count.R
\name{click_count}
\alias{click_count}
\title{click_count}
\usage{
click_count(
  image_path = system.file("example_images", package = "ClickMetrics")
)
}
\description{
This function launch a \pkg{shiny} application in browser to
    count every user click in the image. See section Details for usage.
}
\author{
Felipe de Moraes Kowalski
}
