% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click_poly.R
\name{click_poly}
\alias{click_poly}
\title{click_poly}
\usage{
click_poly(
  image_path = system.file("example_images", package = "ClickMetrics")
)
}
\description{
This function launch a \pkg{shiny} application in browser to
    draw a polygon by clicking on points.
}
\author{
Felipe de Moraes Kowalski
}
