% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_picker.R
\name{color_picker}
\alias{color_picker}
\title{color_picker}
\usage{
color_picker()
}
\description{
This function launch a \pkg{shiny} application in browser to
    detect the hex code and the RGBA of the clicked pixel on the image using javascript.
}
\author{
Felipe de Moraes Kowalski
}
