% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_polygon.R
\name{draw_polygon}
\alias{draw_polygon}
\title{draw_polygon}
\usage{
draw_polygon()
}
\description{
This function launch a \pkg{shiny} application in browser to
    make possible drawing a polygon above a map.
}
\author{
Felipe de Moraes Kowalski
}
